backend/requirements.txt
google-generativeai
flask
flask-cors
fastapi
uvicorn
pydantic